+++ BE/Controllers/StaffController.cs
@@ -1,7 +1,10 @@
 ﻿using ManagerLibrary.Models;
 using ManagerLibrary.Models.DTO;
+using ManagerLibrary.Repository.BookTransactionReponsitory;
+using ManagerLibrary.Repository.Role;
 using ManagerLibrary.Repository.StaffReponsitory;
 using ManagerLibrary.Validation;
+using ManagerLibraryAPI.Repository.ImportBookReponsitory;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -15,11 +18,18 @@ namespace ManagerLibrary.Controllers
     {
         private readonly IStaffReposnitory staffRepository;
         private readonly ValidationStaff validations;
+        private readonly IRoleRepository roleRepository;
+        private readonly IImportBookReponsitory importBookRepository;
+        private readonly IBookTransactionReponsitory bookTransactionRepository;
 
-        public StaffController(IStaffReposnitory staffReposnitory,ValidationStaff validations) 
+        public StaffController(IStaffReposnitory staffReposnitory,ValidationStaff validations,IRoleRepository roleRepository,
+            IImportBookReponsitory importBookReponsitory,IBookTransactionReponsitory bookTransactionReponsitory) 
         {
             this.staffRepository = staffReposnitory;
             this.validations = validations;
+            this.roleRepository=roleRepository;
+            this.importBookRepository = importBookReponsitory;
+            this.bookTransactionRepository = bookTransactionReponsitory;
         }
         [HttpPost]
         
@@ -70,13 +80,16 @@ namespace ManagerLibrary.Controllers
                 }
                 var validationStaff = new ValidationStaffModel
                 {
-                    UserId = string.Empty,
+                    UserId = Id,
                     Phone = model.Phone,
                     Email = model.Email,
                     Address = model.Address,
                     Gender = model.Gender,
                     Avatar = model.Avatar,
-                    RoleId = model.RoleId
+                    RoleId = model.RoleId,
+                    UserName=model.UserName
+                    
+                    
                 };
                 var validationResult = validations.Validate(validationStaff);
                 if (!validationResult.IsValid)
@@ -107,6 +120,12 @@ namespace ManagerLibrary.Controllers
                 {
                     return NotFound();
                 }
+                var import=await importBookRepository.GetAllImportBook(Id);
+                var export =await bookTransactionRepository.GetAllBookTranstion(null,Id, null);
+                if(import.Count>0||export.Count>0)
+                {
+                    return BadRequest(Repository<string>.WithMessage("Không cho xóa vì nhân viên có tạo đơn hàng", 400));
+                }
                 await staffRepository.DeleteStaff(Id);
                 return Ok(Repository<string>.WithMessage("Xóa nhân viên thành công", 200));
             }
@@ -134,13 +153,38 @@ namespace ManagerLibrary.Controllers
                 return BadRequest();
             }
         }
-        [HttpGet]
-        
-        public async Task<IActionResult> GetAllStaff()
+        [HttpGet("Token")]
+        [Authorize(Policy = "IncomeView")]
+
+        public async Task<IActionResult> GetStaffToken()
         {
             try
             {
-                var staff = await staffRepository.GetAllStaff();
+                var staff=await staffRepository.GetStaffToken();
+
+                return Ok(Repository<DTOStaff>.WithData(staff, 200));
+            }
+            catch
+            {
+                return BadRequest();
+            }
+        }
+        [HttpGet]
+        
+        public async Task<IActionResult> GetAllStaff(string?search,string?roleId)
+        {
+            try
+            {
+                if(!string.IsNullOrEmpty(roleId))
+                {
+                    var role = await roleRepository.GetRole(roleId);
+                    if(role == null)
+                    {
+                        return NotFound(Repository<string>.WithMessage("Không tìm thấy roleid",404));
+                    }
+                }
+
+                var staff = await staffRepository.GetAllStaff(search,roleId);
                 
                 return Ok(Repository<List<DTOStaff>>.WithData(staff, 200));
             }
